extract interaction to its own namespace
